+++ utils/kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/validator/TopicsExistenceValidator.scala
@@ -5,16 +5,25 @@ import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.CustomNode
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.api.process.TopicName
 import pl.touk.nussknacker.engine.kafka.UnspecializedTopicName.ToUnspecializedTopicName
+import pl.touk.nussknacker.engine.kafka.validator.TopicsExistenceValidator.TopicValidationType
 
 trait TopicsExistenceValidator extends Serializable {
 
-  final def validateTopic[T <: TopicName](topic: T): Validated[TopicExistenceValidationException[T], T] =
+  final def validateTopic[T <: TopicName: TopicValidationType](
+      topic: T
+  ): Validated[TopicExistenceValidationException[T], T] =
     validateTopics(NonEmptyList.one(topic)).map(_.head)
 
-  def validateTopics[T <: TopicName](
+  def validateTopics[T <: TopicName: TopicValidationType](
       topics: NonEmptyList[T]
   ): Validated[TopicExistenceValidationException[T], NonEmptyList[T]]
 
+}
+
+object TopicsExistenceValidator {
+  sealed trait TopicValidationType[T <: TopicName]
+  implicit case object SourceValidation extends TopicValidationType[TopicName.ForSource]
+  implicit case object SinkValidation   extends TopicValidationType[TopicName.ForSink]
 }
 
 final case class TopicExistenceValidationException[T <: TopicName](topics: NonEmptyList[T])
@@ -28,7 +37,7 @@ object TopicExistenceValidationException {
 
   private def message[T <: TopicName](topics: NonEmptyList[T]): String =
     topics.tail match {
-      case Nil => s"Topic ${topics.head.toUnspecialized.name} doesn't exist"
+      case Nil => s"Topic ${topics.head.toUnspecialized.name} does not exist"
       case _   => s"Topics ${topics.toList.map(_.toUnspecialized.name).mkString(", ")} do not exist"
     }